Rename config to defaultConfigs
